requirements.txt
PyQt5>=5.15
mss>=9.0
openai>=1.0
Pillow>=10.0
keyboard>=0.13
